store: Move PoolCoordinator to its own module
